Perf tests: split correctness counters (PR) from wall-clock (nightly) (doc 0031 M2.4) (#567)
* Perf tests: split correctness counters from wall-clock lanes


* AsyncRendererFilterGroupPerfTestUtils.cc: add missing includes

misc-include-cleaner flagged 6 missing direct includes in this new
file:
- <algorithm> for std::max
- <cstdint> for uint64_t
- <iostream> for std::cerr
- <ratio> for std::milli
- donner/base/EcsRegistry.h for donner::Entity
- donner/svg/compositor/ScopedCompositorHint.h for InteractionHint

* AsyncRendererFilterGroupCorrectness_tests.cc: add <cstdint> for uint64_t

---------
