+++ docs/design_docs/0031-ci_hardening_2026q2.md
@@ -78,7 +78,7 @@ every variant is covered by the default test command.
   - [x] M2.1: Per-config cache slots. Extend `disk-cache` key with a config tag (`-default`, `-asan`, `-geode`). Prevents the asan-fuzzer-evicts-main collision observed pre-Skia-removal. (Subsumes 0029 M1.)
   - [x] M2.2: Extend `tools/cmake/gen_cmakelists.py --check` to optionally `cmake --build` the generated CMake via `--build`, while keeping plain `--check` fast and static. Catches drift that the static validator missed in commits 19d41df8, 398d312b, 89448ad7, a7d682fe.
   - [ ] M2.3: Make `bazel test //...` cover every variant by default. Audit `donner_variant_cc_test` usage; ensure `tiny`, `text-full`, `geode` variants auto-emit under the default test command. Once green, delete `tools/presubmit.sh`. **Blocked on test-surface debt (issue [#566](https://github.com/jwmcglynn/donner/issues/566))** — prototype confirmed the macro path works (471 → 874 test labels, 412 variant wrappers) but running the variants as default gates surfaces 1 tiny-tier + 17 Geode-tier real test failures that must be fixed first.
-  - [ ] M2.4: Introduce `donner_perf_cc_test` macro splitting correctness counters (PR-gate) from wall-clock thresholds (nightly, tagged `perf`). Retires 5 recent threshold-widening hotfixes (8043ad7b, 1f147f2f, 43f42cf7, ab68092b, 8cd89ef7). Absorbs doc 0016 Category 8.
+  - [x] M2.4: Introduce `donner_perf_cc_test` macro splitting correctness counters (PR-gate) from wall-clock thresholds (nightly, tagged `perf`). Retires 5 recent threshold-widening hotfixes (8043ad7b, 1f147f2f, 43f42cf7, ab68092b, 8cd89ef7). Absorbs doc 0016 Category 8.
   - [x] M2.5: Nightly `sanitizers.yml` running ASan+UBSan across `//donner/...`. **Skip-idle**: first job compares `origin/main` HEAD against the last successful `workflow_run` SHA; if unchanged, exit 0 and short-circuit. Monitoring signal only (not PR-blocking).
   - [ ] M2.6: Scheduled Claude triage agent (`/schedule` → `CronCreate`). Runs daily after nightly jobs complete. If the nightly skipped (skip-idle), exit quietly. Otherwise: fetch failure logs, classify (infra vs real bug), file a 🤖-prefixed tracking issue or — for mechanical fixes (missing `target_compatible_with`, threshold widening) — open a PR. Depends on M2.5.
   - [x] M2.7: **Target determinator** via [bazel-diff](https://github.com/Tinder/bazel-diff). A fast `determine-targets` job hashes the PR base + HEAD, emits an affected-target set, and `main.yml`'s build/test jobs run `bazel test <affected>` instead of `//...`. Fallback to `//...` when the PR touches `MODULE.bazel`, `.bazelrc`, `WORKSPACE*`, `build_defs/**`, or `.github/workflows/**`. Typical single-file PR: 15 min → ~3–5 min. (Promotes/replaces old M3.3 "`bazel query rdeps`" sketch — bazel-diff handles `select()` and macro changes more robustly.)
