+++ donner/editor/tests/AsyncRenderer_tests.cc
@@ -1652,138 +1652,6 @@ TEST(AsyncRendererE2ETest, SourcePaneStructurallyEquivalentReparseAvoidsReset) {
   EXPECT_EQ(resetCount, 0u) << "source-pane structurally-equivalent edit took the reset path";
 }
 
-// Regression: dragging a `<g filter=...>` subtree (the editor's flow when
-// the user picks the filter group from the tree view and drags) must go
-// through the compositor's translation-only fast path. Before the subtree-
-// layer fast-path extension, the fast path disqualified ANY layer with
-// `firstEntity != lastEntity` — i.e. anything but a single geometry leaf.
-// That forced the compositor to fall through to
-// `prepareDocumentForRendering`, which rebuilds every RIC, every resolved
-// paint server, and every resolved filter from scratch. On a filter-heavy
-// document the cost is ~100 ms per drag frame, which the user reported as
-// "really laggy" and "I can't select any other elements" (the async
-// renderer stays saturated long enough that the UI-thread's `isBusy()`
-// check never clears).
-//
-// The test measures the drag-frame latency for a filter-group target and
-// asserts a tight budget. Without the subtree fast-path the test blows
-// past it by an order of magnitude.
-TEST(AsyncRendererE2ETest, DraggingFilterGroupSubtreeHitsTranslationFastPath) {
-  std::ifstream splashStream("donner_splash.svg");
-  if (!splashStream.is_open()) {
-    GTEST_SKIP() << "donner_splash.svg not found in runfiles";
-  }
-  std::ostringstream splashBuf;
-  splashBuf << splashStream.rdbuf();
-  const std::string splashSource = splashBuf.str();
-  ASSERT_FALSE(splashSource.empty());
-
-  AsyncSVGDocument asyncDoc;
-  ASSERT_TRUE(asyncDoc.loadFromString(splashSource));
-  asyncDoc.document().setCanvasSize(892, 512);
-  svg::SVGDocument& document = asyncDoc.document();
-
-  // `#Big_lightning_glow` is a `<g filter="url(#big_lightning_glow_blur)">`
-  // containing 2 paths. This is the EXACT shape the user hits when they
-  // drag a filter group on the splash — a subtree layer whose root has a
-  // filter and whose range spans multiple RICs.
-  auto targetElement = document.querySelector("#Big_lightning_glow");
-  ASSERT_TRUE(targetElement.has_value());
-  const Entity targetEntity = targetElement->entityHandle().entity();
-
-  svg::Renderer renderer;
-  AsyncRenderer asyncRenderer;
-
-  const auto waitForResult = [&]() -> std::optional<RenderResult> {
-    const auto deadline = std::chrono::steady_clock::now() + std::chrono::seconds(30);
-    while (std::chrono::steady_clock::now() < deadline) {
-      auto result = asyncRenderer.pollResult();
-      if (result.has_value()) return result;
-      std::this_thread::sleep_for(std::chrono::milliseconds(1));
-    }
-    return std::nullopt;
-  };
-
-  const auto postRequest = [&](uint64_t version, bool activeDrag) {
-    RenderRequest request;
-    request.renderer = &renderer;
-    request.document = &document;
-    request.version = version;
-    request.documentGeneration = 1;
-    request.selectedEntity = targetEntity;
-    request.dragPreview = RenderRequest::DragPreview{
-        .entity = targetEntity,
-        .interactionKind = activeDrag ? svg::compositor::InteractionHint::ActiveDrag
-                                      : svg::compositor::InteractionHint::Selection,
-    };
-    asyncRenderer.requestRender(request);
-  };
-
-  // Pre-warm: selection-kind promotion warms the layer bitmap. Drag
-  // frames that follow should all re-use this cached bitmap via a
-  // compose-offset update, not re-rasterize the filter.
-  postRequest(/*version=*/1, /*activeDrag=*/false);
-  ASSERT_TRUE(waitForResult().has_value());
-
-  constexpr int kDragFrames = 10;
-  double maxDragFrameMs = 0.0;
-  double sumDragFrameMs = 0.0;
-  for (int i = 0; i < kDragFrames; ++i) {
-    targetElement->cast<svg::SVGGraphicsElement>().setTransform(
-        Transform2d::Translate(Vector2d(static_cast<double>(i + 1) * 3.0, 0.0)));
-    const auto t = std::chrono::steady_clock::now();
-    postRequest(/*version=*/static_cast<uint64_t>(i + 2), /*activeDrag=*/true);
-    auto result = waitForResult();
-    const double frameMs =
-        std::chrono::duration<double, std::milli>(std::chrono::steady_clock::now() - t).count();
-    ASSERT_TRUE(result.has_value()) << "filter-group drag frame " << i << " didn't land";
-    ASSERT_TRUE(result->compositedPreview.has_value())
-        << "filter-group drag frame " << i
-        << " produced no composited preview — compositor fell through to the "
-           "non-composited path and will be slow for every subsequent frame";
-    sumDragFrameMs += frameMs;
-    maxDragFrameMs = std::max(maxDragFrameMs, frameMs);
-  }
-
-  const double avgDragFrameMs = sumDragFrameMs / kDragFrames;
-  const auto counters = asyncRenderer.compositorFastPathCountersForTesting();
-  std::cerr << "[PERF] DraggingFilterGroupSubtree: avg=" << avgDragFrameMs
-            << " ms, max=" << maxDragFrameMs << " ms over " << kDragFrames
-            << " frames; fast=" << counters.fastPathFrames
-            << " slowWithDirty=" << counters.slowPathFramesWithDirty
-            << " noDirty=" << counters.noDirtyFrames << "\n";
-
-  // Primary regression gate: the fast path must actually fire. Pre-subtree-
-  // fix the compositor fell through to `prepareDocumentForRendering` EVERY
-  // drag frame, so `slowPathFramesWithDirty` would track 1:1 with drag
-  // frames. The subtree fast path should absorb all 10 drag frames into
-  // `fastPathFrames`. This check is CPU-speed invariant — unlike the
-  // wall-clock budget it doesn't flake on slow CI runners.
-  //
-  // We tolerate up to 1 slow-path frame because the selection pre-warm
-  // render that precedes the drag loop needs the full prepare path to
-  // materialize the initial cached bitmap. Subsequent drag frames must
-  // all take the fast path.
-  EXPECT_GE(counters.fastPathFrames, static_cast<uint64_t>(kDragFrames))
-      << "filter-group subtree drag is not hitting the translation-only fast "
-         "path — compositor is falling through to prepareDocumentForRendering "
-         "every frame";
-  EXPECT_LE(counters.slowPathFramesWithDirty, 1u)
-      << "more than the pre-warm frame slipped into the slow path";
-
-  // Secondary wall-clock budgets: loose enough for shared GitHub CI
-  // runners (which land 30-60 ms per frame on this shape) but still
-  // tight enough to catch the ~250 ms/frame regression this test exists
-  // to gate. The fast-path counter check above is the real regression
-  // detector; these exist so a runaway cost (e.g. someone reintroduces
-  // a full re-rasterize under the fast path) still trips loudly.
-  EXPECT_LT(avgDragFrameMs, 100.0) << "average filter-group drag frame far above even the "
-                                      "widened CI runner budget — something is doing a "
-                                      "full-document re-render per frame";
-  EXPECT_LT(maxDragFrameMs, 200.0) << "max filter-group drag frame exceeded the widened CI "
-                                      "budget — individual frame is doing full-prepare work";
-}
-
 // Regression: a user closing the editor mid-render (or immediately after the
 // last drag frame) was SIGSEGV'ing inside the compositor's `composeLayers`
 // `drawBitmap` lambda. Root cause: `RenderCoordinator` declared
